==> 1_Tests/0_eda.ipynb <==
## Exploratory Data Analysis of the AWS documentation

**Author:** Cristian C. Velandia C.

**Creation Date:** 2024-03-01

This EDA consists on inspecting the different types of markdown files that can be encountered. This analysis helps to identify and select the most suitable straregy to chunk the data for RAG. 

# Clean file names to avoid spaces on them
import os
for f in os.listdir(r"D:\Documents\GitHub\knowledge_pal_assistant\0_data"):
    r = f.replace(" ","")
    if(r != f):
        os.rename(f,r)

### Check data with markdwon loader

from langchain_community.document_loaders import UnstructuredMarkdownLoader

markdown_path = r"D:\Documents\GitHub\knowledge_pal_assistant\0_data\amazon-sagemaker-toolkits.md"
loader = UnstructuredMarkdownLoader(markdown_path)

data = loader.load()

data

data[0].page_content

markdown_path1 = r"D:\Documents\GitHub\knowledge_pal_assistant\0_data\aws-properties-events-rule-sagemakerpipelineparameter.md"
loader1 = UnstructuredMarkdownLoader(markdown_path1, mode="elements")
data1 = loader1.load()

for d in data1:
    print("-"*100)
    print(d)

## Split Markdown on custom tags

from langchain_text_splitters import MarkdownHeaderTextSplitter
from langchain.document_loaders import TextLoader
from langchain.text_splitter import MarkdownTextSplitter

# just ingest the Markdown file raw
data2 = TextLoader(markdown_path1).load()

headers_to_split_on = [("#", "Header 1"), ("##", "Header 2")]

markdown_splitter = MarkdownHeaderTextSplitter(headers_to_split_on=headers_to_split_on, strip_headers = False)
md_header_splits = markdown_splitter.split_text(str(data2[0].page_content))
md_header_splits

for d in md_header_splits:
    print("-"*100)
    print(d)

---

markdown_path2 = "D:\\Documents\\GitHub\\knowledge_pal_assistant\\0_data\\sagemaker-controls.md"
data3 = TextLoader(markdown_path2).load()

headers_to_split_on = [("#", "Header 1"), ("##", "Header 2")]

markdown_splitter = MarkdownHeaderTextSplitter(headers_to_split_on=headers_to_split_on, strip_headers = False)
md_header_splits = markdown_splitter.split_text(str(data3[0].page_content))


for d in md_header_splits:
    print("-"*100)
    print(d)

---
### Test directory loader to speed up data consumption

from langchain_community.document_loaders import DirectoryLoader

markdown_path2 = "D:\\Documents\GitHub\\knowledge_pal_assistant\\0_data"
data4 = DirectoryLoader(markdown_path2, glob = "*.md", recursive = True,  loader_cls = TextLoader,use_multithreading = True, show_progress = True).load()

data4[0].metadata["source"]

# craeate a hashmap of documents to after add the source to each one of the chunks
final_data = {x.metadata["source"]: x.page_content for x in data4}

final_data[data4[0].metadata["source"]]

headers_to_split_on = [("#", "Header 1"), ("##", "Header 2")]

markdown_splitter = MarkdownHeaderTextSplitter(headers_to_split_on=headers_to_split_on, strip_headers = False)

from langchain.document_transformers import 

md_header_splits = []

for source, docs in final_data.items():
    md_header_splits.extend([{"page_content": chunk.page_content + "\n \'data source =  {0}\'".format(source), "metadata":chunk.metadata} for chunk in markdown_splitter.split_text(docs)])
    break

md_header_splits[0]
