==> 1_Tests/1_prepare_chunk_data.ipynb <==
# Dataset for embeddings Creation

**Author:** Cristian C. Velandia C.

**Creation Date:** 2024-03-02

This notebook aims to create a dataset stored in parquet with the necessary data to create the embeddings and after upsert to pinecone VDB. 

Data is cleaned before the embeddings process

from langchain_community.document_loaders import DirectoryLoader
from langchain_text_splitters import MarkdownHeaderTextSplitter
from langchain.document_loaders import TextLoader
import pandas as pd

---
## Data Loading

markdown_path2 = "D:\\Documents\GitHub\\knowledge_pal_assistant\\0_data"
# Define the directory loader, be carefull with use_multithreading as it will output the files in different order every time
data4 = DirectoryLoader(markdown_path2, glob = "*.md", recursive = True, loader_cls = TextLoader, use_multithreading = True, show_progress = True, sample_seed = 1).load()

data4[0].metadata["source"]

len(data4)

---
Data Chunking and cleaning

# Create a hashmap of documents to after add the source to each one of the chunks
final_data = {x.metadata["source"]: x.page_content for x in data4}

len(final_data)

#Check results
final_data[data4[0].metadata["source"]]

# Define over which headers the data will be split and chunked
headers_to_split_on = [("#", "Header 1"), ("##", "Header 2")]
# Define the splitter object
markdown_splitter = MarkdownHeaderTextSplitter(headers_to_split_on=headers_to_split_on, strip_headers = False)

md_header_splits = [] #Define an empty list to extend with all the corpus

# Go through all sources and doocuments
for id, (source, docs) in enumerate(final_data.items()):
    md_header_splits.extend([{"id": id, "chunk": i, "page_content": chunk.page_content + "\n \'data source =  {0}\'".format(source), "source" : source , "metadata" : chunk.metadata} for i, chunk in enumerate(markdown_splitter.split_text(docs))])

len(md_header_splits)

md_header_splits[0]

#Create a dataframe from the list of dicts to store data, deduplicate, and clean the chunks
final_df = pd.DataFrame(md_header_splits)

# OPTIONAL CLEAN STOPWORDS 
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
#from nltk import download
#download('stopwords')

stop_words = set(stopwords.words('english'))
#generate tokens
final_df["tokens"] = final_df["page_content"].apply(lambda x: word_tokenize(x))
#Remove Stop words
final_df["nostopw_page_content"] = final_df["tokens"].apply(lambda x: " ".join([w for w in x if not w.lower() in stop_words]))

final_df.head()

#add chunk and source to meta 
final_df["metadata"] = final_df.loc[:, ["chunk", "source", "metadata", "page_content"]].apply(lambda x:  {**x["metadata"],**{"chunk" : x["chunk"], "source" : x["source"], "text" : x["page_content"]}}, axis=1)
final_df["metadata"].head()


final_df["metadata"][4]

final_df[final_df["id"] == 0]

#generate final id
final_df["id"] = final_df["id"].astype('str') + "-" + final_df["chunk"].astype('str')

#Drop unnecessary columns
final_df.drop(columns = ["chunk", "source"], inplace = True)

final_df.shape

final_df.head()

save_folder = "D:\\Documents\GitHub\\knowledge_pal_assistant\\2_outputs\\"
final_df.to_parquet(save_folder + "chunks.parquet", index = False, engine = "pyarrow", compression= "brotli")
