==> 1_Tests/2_1_test_embeddings_openai.ipynb <==
# Generate Embeddings for text
Author: Cristian Velandia

Creation Date: 2024-03-02

Create embeddings for the vector DB using openai embeddings API

from openai import OpenAI
from tqdm.auto import tqdm
from time import sleep
import pandas as pd
import json

### Load Dataset

data_path = r"D:\Documents\GitHub\knowledge_pal_assistant\2_outputs\chunks.parquet"
chunks = pd.read_parquet(data_path, engine = "pyarrow")

chunks.head()

chunks.shape

chunks["token_length"] = chunks["tokens"].apply(lambda x: len(x))
chunks.reset_index(drop = False, inplace = True)

### Connect to OpenAI

# get api keys
creds =json.load(open('personal_creds.json'))

OPENAI_API_KEY = creds["OPENAI_API_KEY"]

openai_client = OpenAI(api_key=OPENAI_API_KEY)

### Generate Embeddings

# To embed multiple inputs in a single request, pass an array of strings or array of token arrays. The input must not exceed the max input tokens for the model (8192 tokens for text-embedding-ada-002). API limits 200 CPD, 150000 TPM free tier, 3 requests per minute
#chunks["vectors"] = chunks["page_content"].apply(lambda x: openai_client.embeddings.create(input = x, model = "text-embedding-ada-002"))


#Create Batches and request embedings
corpus_size = chunks.shape[0]
counter = 1
batch_size = 5
valid_batch = 0
vect_df = pd.DataFrame()
toggle = False
max_calls = 100 

while corpus_size > 0:
    batch = chunks[(chunks["index"] >= batch_size*(counter-1)) & (chunks["index"] < batch_size*counter)]
    corpus_size -= batch.shape[0]
    text_list = batch["page_content"].to_list()
    total_tokens = batch["token_length"].sum()
    
    if total_tokens <= 8192:
        valid_batch += 1
        print("ammount of valid batches: ", valid_batch)
        valid_ids = batch["id"].tolist()

        #Make sure to not do more than 3 requests per minute, try 10 times 20 secs spaced
        for attempt in range(10):
            try:
                #Call Embeddings API
                embeddings = openai_client.embeddings.create(input = text_list, model = "text-embedding-ada-002",)
            except Exception as error:
                print(type(error).__name__, "-", error, "now sleep")
                if "RPD" in str(error): #handle max requests per day
                    toggle = True
                    break
                else: #handle max requests per minute 
                    sleep(60)
            else:
                print("ok")
                break
        
        vect_df = pd.concat([vect_df, pd.DataFrame.from_dict(dict(zip(valid_ids, [[e.embedding] for e in embeddings.data])), orient='index', columns=["vect"])])

    if toggle or valid_batch >= 100:
        print("Terminating due to max requests exceeded")
        break

        
    
    print("count {}, batch {}, tokens in batch {}, remaining rows {}, vectors generated {}".format(counter, batch.shape[0], total_tokens, corpus_size, vect_df.shape[0]))
        
    counter += 1

valid_batch

vectorized_chunks = chunks.merge(vect_df.reset_index(drop=False), how="inner", left_on="id", right_on="index")

vectorized_chunks.head()

save_folder = "D:\\Documents\GitHub\\knowledge_pal_assistant\\2_outputs\\"
vectorized_chunks.to_parquet(save_folder + "vectors.parquet", index = False, engine = "pyarrow", compression= "brotli")
