==> 1_Tests/2_2_test_embeddings_llama.ipynb <==
# Generate Embeddings for text
Author: Cristian Velandia

Creation Date: 2024-03-03

Create embeddings for the vector DB using Metas Llama cpp quntize embeddings model

from langchain.embeddings import LlamaCppEmbeddings
from llama_cpp import Llama
import pandas as pd
from huggingface_hub import hf_hub_download

### Load Dataset

data_path = r"D:\Documents\GitHub\knowledge_pal_assistant\2_outputs\chunks.parquet"
chunks = pd.read_parquet(data_path, engine = "pyarrow")

chunks.head()

chunks.shape

chunks["token_length"] = chunks["tokens"].apply(lambda x: len(x))
chunks.reset_index(drop = False, inplace = True)

chunks["token_length"].value_counts().sort_index(ascending = False)

### Setup LLama Embeddings model

models_folder = ".\\hf_models"
model_name = "TheBloke/Llama-2-7B-Chat-GGML"
model_filename = "llama-2-7b-chat.ggmlv3.q4_K_M.bin"

#Download the model if necessary
file_path = hf_hub_download(repo_id = model_name, filename = model_filename, local_dir = models_folder,)

llama_model_path = f"{models_folder}\\{model_filename}"
embeddings = LlamaCppEmbeddings(model_path = llama_model_path, n_ctx = 4096, n_gpu_layers = 5) # Instance model with GPU n_gpu_layers = 30 works slower than just processor, n_gpu_layers = 5 optimal 

### Generate Embeddings

documents = chunks[chunks["token_length"] <= 4096]["page_content"].to_list()
len(documents)

test_string = documents[11]
embs = embeddings.embed_query(test_string)

len(embs)

save_folder = "D:\\Documents\GitHub\\knowledge_pal_assistant\\2_outputs\\"
corpus_size = chunks.shape[0]
batch_size = 50
counter = 1
dmb_df = pd.DataFrame()

while corpus_size > 0:
    print(corpus_size)

    batch = chunks[(chunks["index"] >= batch_size*(counter-1)) & (chunks["index"] < batch_size*counter)][["id", "page_content"]]
    documents = batch["page_content"].to_list()

    # Embed texts
    aws_embeddings = embeddings.embed_documents(documents)

    # Checkpoint
    batch_ids = batch["id"].tolist()
    dmb_df = pd.concat([dmb_df, pd.DataFrame.from_dict(dict(zip(batch_ids, [[e] for e in aws_embeddings])), orient='index', columns=["vect"])])
    vectorized_chunks = chunks.merge(dmb_df.reset_index(drop=False), how="inner", left_on="id", right_on="index")
    vectorized_chunks.to_parquet(save_folder + "vectors_llama.parquet", index = False, engine = "pyarrow", compression= "brotli")

    # Control Parameters
    corpus_size -= batch.shape[0]
    counter += 1   

vectorized_chunks.shape
