==> 1_Tests/4_rag_pinecone_llama.ipynb <==
# Retrieval Augmented Generation
Author: Cristian Velandia

Creation Date: 2024-03-02

RAG Using PINECONE vector DB, OpenAI chat gpt 3.5 and openai embeddings

from pinecone import Pinecone, ServerlessSpec
from llama_cpp import Llama
from tqdm.auto import tqdm
import pandas as pd
import json

from langchain.embeddings import LlamaCppEmbeddings
from llama_cpp import Llama


# get api keys
creds =json.load(open('personal_creds.json'))

PINECONE_API_KEY = creds["PINECONE_API_KEY"]

## Setup Pinecone
### Log into API

pinecone = Pinecone(api_key=PINECONE_API_KEY)

# Dfine unique and readable index name
INDEX_NAME = 'aws-docs-vdb-index'

### Create Index for Vector storage

# Check if index already exists and deletes it
if INDEX_NAME in [index.name for index in pinecone.list_indexes()]:
  pinecone.delete_index(INDEX_NAME)

# Recreates index, after creation it is visible at the consle
pinecone.create_index(name = INDEX_NAME, dimension = 4096, metric = 'cosine',  spec = ServerlessSpec(cloud = 'aws', region = 'us-west-2')) 

#Create index object
index = pinecone.Index(INDEX_NAME)

### Load Previously Created Dataset
Here we load the dataset with the crpus and vectors to upload them after into pinecone (upsert)

data_path = r"D:\Documents\GitHub\knowledge_pal_assistant\2_outputs\vectors_llama.parquet"
vectors = pd.read_parquet(data_path, engine = "pyarrow")

vectors.head()

#clean meta, do not include headers
def pop_keys(d, key):
    tmp = d.copy()
    tmp.pop(key)
    return tmp

vectors["metadata"] = vectors["metadata"].apply(pop_keys, args = (["Header 1"])) 
vectors["metadata"] = vectors["metadata"].apply(pop_keys, args = (["Header 2"])) 

vectors["metadata"][0]

### Upsert embeddings to Pinecone 

# Create list for batching the upload
prepped = []

# Iterate through data
for i, row in tqdm(vectors.iterrows(), total=vectors.shape[0]):

    prepped.append({'id' : row['id'], 'values' : row['vect'], 'metadata' : row['metadata']})

    if len(prepped) >= 100:
        index.upsert(prepped) #Upsert a batch of 200 vectors
        prepped = []


#Descripe uploaded index
index.describe_index_stats()

### Augment Llama Chat Queries 
Trhough this section the embeddings and calls to the LLM are performed, after, the prompts are built and tested. Here we will test the RAG with the test questions provided 

models_folder = ".\\hf_models"
model_name = "TheBloke/Llama-2-7B-Chat-GGML"
model_filename = "llama-2-7b-chat.ggmlv3.q4_K_M.bin"
llama_model_path = f"{models_folder}\\{model_filename}"

llm_model = Llama(llama_model_path, n_ctx = 4096)
emb_model = LlamaCppEmbeddings(model_path = llama_model_path, n_ctx = 4096, n_gpu_layers = 5) # Instance model with GPU 

# First step, get all the relevant infromation from the vector DB (this will be the question that the user will ask)
query = "What are all AWS regions where SageMaker is available?"

embs = emb_model.embed_query(query)

res = index.query(vector = embs, top_k = 3, include_metadata = True)

context = [r['metadata']['text'] for r in res['matches']]

print('\n'.join(context)) #Visualize output


BOS, EOS = "<s>", "</s>"
B_INST, E_INST = "[INST]", "[/INST]"
B_SYS, E_SYS = "<<SYS>>\n", "\n<</SYS>>\n\n"
SYSTEM_BEHAVIOR = """You are "Knowledge pal", an assistant for developers. Your role consists on answering questions about cloud services, coding in different languages, and provide a detailed response everytime. You will provide the source of the answer you are giving and related sources. Those sources can be a path or URL. Answer the question by combining your knowledge with the context provided. If information is not clear or the context is not enough to give an answer, tell the user that you don't have the answer. \n\n """

message = B_INST + B_SYS + SYSTEM_BEHAVIOR + "Context: " + '\n'.join(context) + E_SYS + "Question: " + query + " Answer: " + E_INST
print(message)

knowledge_pal_response_gen = llm_model(prompt = message, stream = False, temperature = 0.2, top_p = 0.6, max_tokens = 512, stop = None)

print(knowledge_pal_response_gen["choices"][0]["text"])

# Code for streaming=True
knowledge_pal_response = "".join([c["choices"][0]["text"] for c in knowledge_pal_response_gen])
print('-' * 80)
print(knowledge_pal_response)

# Chat Helper for messaging
